Create molecule.converter module for RDKit and OB conversion
Move to/from RDKit/OB methods to new module
Update affected imports
